project_june/requirements.txt
requests
pandas
psycopg2-binary 